% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the Company Data Warehouse (DVH)}
\usage{
db_connect(database = "DWHI", uid = NULL, dwhi_cred = c("x", "y", "z"))
}
\arguments{
\item{database}{Character.}

\item{uid}{Character. Your user identification.}

\item{dwhi_cred}{Character vector.}
}
\value{
A DBIConnection object.
}
\description{
Establishes an ODBC connection to the data warehouse environments.
}
