% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hent_data.R
\name{hent_data}
\alias{hent_data}
\title{Fetch Data from DVH}
\usage{
hent_data(connection, query = "")
}
\arguments{
\item{connection}{A DBIConnection object created by \code{db_connect()}.}

\item{query}{Character string containing the SQL query.}
}
\value{
A data frame.
}
\description{
Executes a query against the established database connection.
}
